+++ changelog.md
@@ -1,3 +1,9 @@
+Version 1.2 – 9 March 2015
+--------------------------
+
+* Updated for the Terraria Server API version 1.17.
+* I now use the RegionDeleted hook to inform users immediately and descriptively when a region they are viewing is deleted. It turns out I can't use RegionEntered and RegionLeft, because they don't work when regions overlap.
+
 Version 1.1 – 21 July 2014
 --------------------------